Purchase and sales formatted and view data in sales

+++ src/com/cherp/entities/Purchase.java
@@ -12,185 +12,143 @@ public class Purchase {
 	private String cleaner2;
 	private String company;
 	private String location;
-	private long outstanding;
+	private int outstanding;
 	private long challanNo;
-	private long rent;
+	private int rent;
 	private String product;
 	private int pieces;
-	private long kg;
-	private long rate;
-	private long amount;
-	private long avgWeight;
+	private int kg;
+	private int rate;
+	private int amount;
+	private Double avgWeight;
 	private int status;
 	private String combinePurchaseToggle;
-
-	
 	public int getId() {
 		return id;
 	}
-
 	public void setId(int id) {
 		this.id = id;
 	}
-
 	public int getPurchaseId() {
 		return purchaseId;
 	}
-
 	public void setPurchaseId(int purchaseId) {
 		this.purchaseId = purchaseId;
 	}
-
 	public String getDate() {
 		return date;
 	}
-
 	public void setDate(String date) {
 		this.date = date;
 	}
-
 	public String getVanName() {
 		return vanName;
 	}
-
 	public void setVanName(String vanName) {
 		this.vanName = vanName;
 	}
-
 	public String getDriver1() {
 		return driver1;
 	}
-
 	public void setDriver1(String driver1) {
 		this.driver1 = driver1;
 	}
-
 	public String getDriver2() {
 		return driver2;
 	}
-
 	public void setDriver2(String driver2) {
 		this.driver2 = driver2;
 	}
-
 	public String getCleaner1() {
 		return cleaner1;
 	}
-
 	public void setCleaner1(String cleaner1) {
 		this.cleaner1 = cleaner1;
 	}
-
 	public String getCleaner2() {
 		return cleaner2;
 	}
-
 	public void setCleaner2(String cleaner2) {
 		this.cleaner2 = cleaner2;
 	}
-
 	public String getCompany() {
 		return company;
 	}
-
 	public void setCompany(String company) {
 		this.company = company;
 	}
-
 	public String getLocation() {
 		return location;
 	}
-
 	public void setLocation(String location) {
 		this.location = location;
 	}
-
-	public long getOutstanding() {
+	public int getOutstanding() {
 		return outstanding;
 	}
-
-	public void setOutstanding(long outstanding) {
+	public void setOutstanding(int outstanding) {
 		this.outstanding = outstanding;
 	}
-
 	public long getChallanNo() {
 		return challanNo;
 	}
-
 	public void setChallanNo(long challanNo) {
 		this.challanNo = challanNo;
 	}
-
-	public long getRent() {
+	public int getRent() {
 		return rent;
 	}
-
-	public void setRent(long rent) {
+	public void setRent(int rent) {
 		this.rent = rent;
 	}
-
 	public String getProduct() {
 		return product;
 	}
-
 	public void setProduct(String product) {
 		this.product = product;
 	}
-
 	public int getPieces() {
 		return pieces;
 	}
-
 	public void setPieces(int pieces) {
 		this.pieces = pieces;
 	}
-
-	public long getKg() {
+	public int getKg() {
 		return kg;
 	}
-
-	public void setKg(long kg) {
+	public void setKg(int kg) {
 		this.kg = kg;
 	}
-
-	public long getRate() {
+	public int getRate() {
 		return rate;
 	}
-
-	public void setRate(long rate) {
+	public void setRate(int rate) {
 		this.rate = rate;
 	}
-
-	public long getAmount() {
+	public int getAmount() {
 		return amount;
 	}
-
-	public void setAmount(long amount) {
+	public void setAmount(int amount) {
 		this.amount = amount;
 	}
-
-	public long getAvgWeight() {
+	public Double getAvgWeight() {
 		return avgWeight;
 	}
-
-	public void setAvgWeight(long avgWeight) {
+	public void setAvgWeight(Double avgWeight) {
 		this.avgWeight = avgWeight;
 	}
-
 	public int getStatus() {
 		return status;
 	}
-
 	public void setStatus(int status) {
 		this.status = status;
 	}
-
 	public String getCombinePurchaseToggle() {
 		return combinePurchaseToggle;
 	}
-
 	public void setCombinePurchaseToggle(String combinePurchaseToggle) {
 		this.combinePurchaseToggle = combinePurchaseToggle;
 	}
-
+	
+	
 }
